Исправление ошибки в методе CalculatorPresenter::formatText. Теперь положение курсора внутри дробных чисел находится корректно

+++ app/src/main/java/GeekBrians/Slava_5655380/Calculator/Calculator.java
@@ -5,7 +5,7 @@ import java.math.MathContext;
 import java.math.RoundingMode;
 
 public class Calculator implements BinaryOperation {
-    private int precision = 16;
+    private int precision = 13;
     private RoundingMode roundingMode = RoundingMode.HALF_UP;
 
     @Override
